Add files via upload
One day I wanted to type things backwards on my work instant messaging app, so i created a script to help me do it. Then I was interested in "reversing" whole texts, such as the U.S. Declaration of Independence. 

Who knows if I'll continue with this or not? Could be fun